React to code review
This incorporates changes suggested by Kevin Barabash.

In particular, command line arguments are now handled
as named options, not just by position, making this far more flexible.
Contrary to the review, nomnom is used instead of minimist,
which generates a nice online help and this improves the documentation.

This commit also applies lint to the new script, and deals with its complaints.